splits transform.rs into multiple submodules
